Improve conversion script
