+++ src/transformers/models/blip_2/test.py
@@ -1,5 +1,6 @@
 from transformers import Blip2Config, Blip2ForConditionalGeneration
 
+
 config = Blip2Config()
 
 model = Blip2ForConditionalGeneration(config)
